+++ sdk/base/nostd/set.hpp
@@ -78,10 +78,17 @@ class t_set_runtime {
         return *this;
     }
     bool contains(const decayed_t& value) { return _target->contains(value); }
+
     template <typename U = decayed_t>
     typename std::enable_if<std::is_same<U, std::string>::value, bool>::type  //
     match(const decayed_t& value) {
         return _target->match(value);
+    }
+
+    template <typename U = decayed_t>
+    typename std::enable_if<std::is_same<U, std::string>::value, bool>::type  //
+    regex(const decayed_t& value) {
+        return _target->regex(value);
     }
 
     template <typename U = decayed_t>